requirements.txt
transformers
torch
torchvision
pillow
streamlit
requests
